c314: extract BenchmarksGating tab cluster (DeepGateSection + NeuralTopicComparisonSection + BayesianHdiSection) to own file (#441 P0 2.2)
Benchmarks.tsx 2645 LOC -> 2120 LOC (-525 LOC). Benchmarks chunk 61.92 KB -> 49.11 KB (-12.8 KB).
New chunk: BenchmarksGating-Cw7dZZ7t.js 13.39 KB.
typecheck clean, 44/44 tests pass.
